requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
